[REFACTOR]: migrate supplier_management module to supplier package with updated structure
