+++ src/main/java/id/ac/ui/cs/advprog/buildingstore/supplier/model/PurchaseTransaction.java
@@ -1,0 +1,29 @@
+package id.ac.ui.cs.advprog.buildingstore.supplier.model;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class PurchaseTransaction {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String productName;
+    private int quantity;
+    private BigDecimal totalPrice;
+    private LocalDateTime date;
+
+    @ManyToOne
+    @JoinColumn(name = "supplier_id", nullable = false)
+    private Supplier supplier;
+}
