+++ src/main/java/id/ac/ui/cs/advprog/buildingstore/supplier/model/Supplier.java
@@ -1,0 +1,27 @@
+package id.ac.ui.cs.advprog.buildingstore.supplier.model;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.util.List;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class Supplier {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String name;
+    private String address;
+    private String contact;
+    @Enumerated(EnumType.STRING)
+    private SupplierCategory category;
+    @OneToMany(mappedBy = "supplier", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<PurchaseTransaction> transactions;
+}
