+++ src/main/java/id/ac/ui/cs/advprog/buildingstore/supplier/model/SupplierCategory.java
@@ -1,0 +1,10 @@
+package id.ac.ui.cs.advprog.buildingstore.supplier.model;
+
+public enum SupplierCategory {
+    SEMEN,
+    KAYU,
+    BESI_BAJA,
+    LISTRIK,
+    PLUMBING,
+    ATAP_DINDING
+}
