--- src/main/java/id/ac/ui/cs/advprog/buildingstore/supplier_management/model/SupplierCategory.java
@@ -1,10 +1,0 @@
-package id.ac.ui.cs.advprog.buildingstore.supplier_management.model;
-
-public enum SupplierCategory {
-    SEMEN,
-    KAYU,
-    BESI_BAJA,
-    LISTRIK,
-    PLUMBING,
-    ATAP_DINDING
-}
