--- src/main/resources/templates/admin/add_supplier.html
@@ -1,72 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8" />
-    <title>Add Supplier</title>
-    <script src="https://cdn.tailwindcss.com"></script>
-</head>
-<body class="bg-gray-100 min-h-screen flex">
-
-<!-- Sidebar -->
-<div th:replace="fragments/navbar :: navbar"></div>
-
-<!-- Main Content -->
-<main class="flex-1 p-10">
-    <div class="max-w-xl mx-auto bg-white rounded-3xl shadow-xl p-8">
-        <h1 class="text-2xl font-bold text-gray-900 mb-6">Add New Supplier</h1>
-
-        <form th:action="@{/supplier/add}" method="post" th:object="${supplierDTO}" class="space-y-6">
-            <input type="hidden" th:name="${_csrf.parameterName}" th:value="${_csrf.token}" />
-
-            <!-- Name -->
-            <div>
-                <label class="block mb-1 font-medium text-gray-700">Name</label>
-                <input type="text" th:field="*{name}" required
-                       class="w-full rounded-xl border px-4 py-3 text-gray-900 placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-blue-600"
-                       placeholder="Supplier name" />
-                <p th:if="${#fields.hasErrors('name')}" th:errors="*{name}" class="text-red-600 text-sm mt-1"></p>
-            </div>
-
-            <!-- Address -->
-            <div>
-                <label class="block mb-1 font-medium text-gray-700">Address</label>
-                <input type="text" th:field="*{address}" required
-                       class="w-full rounded-xl border px-4 py-3 text-gray-900 focus:outline-none focus:ring-2 focus:ring-blue-600"
-                       placeholder="Supplier address" />
-                <p th:if="${#fields.hasErrors('address')}" th:errors="*{address}" class="text-red-600 text-sm mt-1"></p>
-            </div>
-
-            <!-- Contact -->
-            <div>
-                <label class="block mb-1 font-medium text-gray-700">Contact</label>
-                <input type="text" th:field="*{contact}" required
-                       class="w-full rounded-xl border px-4 py-3 text-gray-900 focus:outline-none focus:ring-2 focus:ring-blue-600"
-                       placeholder="Phone or email" />
-                <p th:if="${#fields.hasErrors('contact')}" th:errors="*{contact}" class="text-red-600 text-sm mt-1"></p>
-            </div>
-
-            <!-- Category -->
-            <div>
-                <label class="block mb-1 font-medium text-gray-700">Category</label>
-                <select th:field="*{category}" required
-                        class="w-full rounded-xl border px-4 py-3 text-gray-900 focus:outline-none focus:ring-2 focus:ring-blue-600">
-                    <option value="" disabled selected>Select Category</option>
-                    <option th:each="cat : ${T(id.ac.ui.cs.advprog.buildingstore.supplier_management.model.SupplierCategory).values()}"
-                            th:value="${cat}" th:text="${cat.name()}"></option>
-                </select>
-                <p th:if="${#fields.hasErrors('category')}" th:errors="*{category}" class="text-red-600 text-sm mt-1"></p>
-            </div>
-
-            <!-- Submit -->
-            <div class="text-right">
-                <button type="submit"
-                        class="bg-blue-600 text-white px-6 py-3 rounded-xl hover:bg-blue-700 transition">
-                    Save Supplier
-                </button>
-            </div>
-        </form>
-    </div>
-</main>
-
-</body>
-</html>
