+++ src/main/resources/templates/supplier/add_transaction.html
