+++ src/main/resources/templates/supplier/supplier_list.html
