+++ src/main/resources/templates/supplier/supplier_transactions.html
